Creates julia environment for free_convection problem
