+++ free_convection/Project.toml
@@ -1,0 +1,10 @@
+name = "free_convection"
+uuid = "8bef6888-3ab9-11e9-188b-1bf153246cf9"
+authors = ["Gregory LeClaire Wagner <wagner.greg@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
+Oceananigans = "6e00b360-fae3-11e8-3922-3d7cb9e50aad"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
